consolidate dataset saving
