+++ yourbench/pipeline/answer_generation.py
@@ -62,7 +62,7 @@ from loguru import logger
 from datasets import Dataset
 
 from yourbench.utils.dataset_engine import smart_load_dataset
-from yourbench.utils.saving_engine import save_dataset
+from yourbench.utils.dataset_engine import save_dataset
 from yourbench.utils.inference_engine import InferenceCall, run_inference
 from yourbench.utils.prompts import (
     ZEROSHOT_QA_USER_PROMPT,
