+++ yourbench/pipeline/chunking.py
@@ -46,7 +46,7 @@ from transformers import AutoTokenizer, AutoModel
 import torch.nn.functional as F
 
 from yourbench.utils.dataset_engine import smart_load_dataset
-from yourbench.utils.saving_engine import save_dataset
+from yourbench.utils.dataset_engine import save_dataset
 
 import evaluate
 _perplexity_metric = None
