+++ yourbench/pipeline/single_shot_question_generation.py
@@ -23,7 +23,7 @@ from loguru import logger
 from datasets import Dataset
 
 from yourbench.utils.dataset_engine import smart_load_dataset
-from yourbench.utils.saving_engine import save_dataset
+from yourbench.utils.dataset_engine import save_dataset
 from yourbench.utils.inference_engine import InferenceCall, run_inference
 from yourbench.utils.prompts import (
     QUESTION_GENERATION_SYSTEM_PROMPT,
