+++ yourbench/pipeline/summarization.py
@@ -16,7 +16,7 @@ _bertscore = evaluate.load("bertscore")
 
 from yourbench.utils.inference_engine import run_inference, InferenceCall
 from yourbench.utils.prompts import SUMMARIZATION_USER_PROMPT
-from yourbench.utils.saving_engine import save_dataset
+from yourbench.utils.dataset_engine import save_dataset
 from yourbench.utils.dataset_engine import smart_load_dataset
 from yourbench.utils.parsing_engine import extract_content_from_xml_tags
